gism ... /all was never running for source windowsfeatures
Issue here was minor version was always required to be > 2, so now that we're at windows 10.0.x /all is not being added, even though this logic really only wanted to make sure the version was >= 6.2

Added a check to see if the version is greater than 6 entirely, if not, only then does it check if its version 6 and at least minor ver. 6.2.

Not calling /all as part of gism meant that dependencies for windows features were never installed as part of the installation.